Move interface yiisoft/db to yiisoft/db/contracts.
