--- src/ConnectionInterface.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Yiisoft\Db;
-
-interface ConnectionInterface
-{
-}
