+++ src/Contracts/ConnectionInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Yiisoft\Db\Contracts;
+
+interface ConnectionInterface
+{
+    public function open();
+
+    public function close();
+}
