+++ src/Contracts/ExpressionInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yiisoft\Db;
+namespace Yiisoft\Db\Contracts;
 
 /**
  * Interface ExpressionInterface should be used to mark classes, that should be built
